TankDrive Update
I've updated the TankDrive subsystem to take in strings "FT" and "LR" for modes. Here is what this does:

With "FT", the code maps the X-axis and Y-axis controllers with front (both sides moving in the same direction -> forward movement) and turn (both sides moving in opposite directions -> turning movement).

With "LR", the X-axis and Y-axis are mapped to left-side movement and right-side movement respectively.

As said before, let me know if I need to improve anything!