added the ability to know where we are based on april tags
The next things I want to implement is to figure out to how put in that information into the path planner also the swerve subsystem. But I think that this is a good base line. We also still need to be able align with the AprilTags when we want to shoot and climb